Changing the name of the module from invkin to pickit
As in Pick it - a much much dumber version of Move it :)
